+++ tests/args.py
@@ -11,5 +11,8 @@ class Args:
         self.zig_zag_rev = False
         self.diagonal = False
         self.all = False
+        self.symbols = False
+        self.digits = False
+        self.file = False
         self.minlen = 6
         self.maxlen = 8